# pkg/stringid: optimize performance